[Bitmex] Rename test classes
